_requirements.txt
aider-chat==0.63.2
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.4.1
anaconda-anon-usage @ file:///C:/b/abs_c3w_h1zzjg/croot/anaconda-anon-usage_1710965204622/work
annotated-types==0.7.0
anyio==4.6.2.post1
archspec @ file:///croot/archspec_1709217642129/work
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
boltons @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/boltons_1699480450092/work
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools==5.5.0
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi @ file:///C:/b/abs_90yq4lmu83/croot/cffi_1726856448345/work
charset-normalizer==3.4.0
click==8.1.7
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
conda @ file:///C:/b/abs_85jnuwc__u/croot/conda_1729193917673/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///C:/b/abs_831_pypuuz/croot/conda-package-handling_1718138287790/work
conda_package_streaming @ file:///C:/b/abs_fe6zb0_j8_/croot/conda-package-streaming_1718136098642/work
ConfigArgParse==1.7
cryptography @ file:///C:/b/abs_35g500qir4/croot/cryptography_1724940575116/work
diff-match-patch==20241021
diskcache==5.6.3
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
filelock==3.16.1
flake8==7.1.1
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
grep-ast==0.3.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.7.1
json5==0.9.28
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
libmambapy @ file:///C:/b/abs_aam1usey89/croot/mamba-split_1725563515831/work/libmambapy
linkify-it-py==2.0.3
litellm==1.52.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
menuinst @ file:///C:/b/abs_76jwd47ck1/croot/menuinst_1723567618942/work
mixpanel==4.10.1
monotonic==1.6
multidict==6.1.0
narwhals==1.13.5
networkx==3.2.1
numpy==1.26.4
openai==1.54.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/platformdirs_1701797392447/work
pluggy @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pluggy_1699472504117/work
posthog==3.7.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pyarrow==18.0.0
pycodestyle==2.12.1
pycosat @ file:///C:/b/abs_5csdern___/croot/pycosat_1714513102923/work
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pypandoc==1.14
pyperclip==1.9.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
rich==13.9.4
rpds-py==0.21.0
ruamel.yaml @ file:///C:/b/abs_0cunwx_ww6/croot/ruamel.yaml_1727980181547/work
ruamel.yaml.clib @ file:///C:/b/abs_5fk8zi6n09/croot/ruamel.yaml.clib_1727769837359/work
scipy==1.13.1
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
streamlit==1.40.1
tenacity==8.5.0
textual==1.0.0
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
truststore @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/truststore_1701881385424/work
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
watchdog==4.0.2
wcwidth==0.2.13
wheel==0.44.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
yarl==1.17.1
zipp==3.21.0
zstandard @ file:///C:/b/abs_82dff2ths3/croot/zstandard_1728569207191/work
